chore: split FieldSet out of FederationKey
