+++ engine/crates/engine/src/registry/requires.rs
@@ -1,0 +1,4 @@
+//! Fields can declare that they "require" other fields.
+//!
+//! This is usually used on a field with a custom resolver where that custom resolver
+//! needs some context from the containing type in order to do its work.
